+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-## 0.1.1 - Unreleased
+## 0.1.1 - 2024-02-18
 
 BUGS:
  * fix guest stats structure name
+
+FEATURES:
+ * add new function to query minidisks
 
 
 ## 0.1.0 - 2023-12-31